feature/CP-9006-ios-sdk-automatically-show-app-banner-when-app-is-open-in-the-foreground-and-handle-notification-received-is-called-via-silent-push (#270)
* feature/CP-9006-ios-sdk-automatically-show-app-banner-when-app-is-open-in-the-foreground-and-handle-notification-received-is-called-via-silent-push

Implemented functionality of app banners should be displayed while the application is in foreground mode and the push notification type is a silent push.

* feature/CP-9006-ios-sdk-automatically-show-app-banner-when-app-is-open-in-the-foreground-and-handle-notification-received-is-called-via-silent-push

Implemented functionality of app banners should be displayed while the application is in foreground mode and the push notification type is a silent push.

* feature/CP-9006-ios-sdk-automatically-show-app-banner-when-app-is-open-in-the-foreground-and-handle-notification-received-is-called-via-silent-push

Implemented functionality of app banners should be displayed while the application is in foreground mode and the push notification type is a silent push.