feature/CP-9383-ios-remove-notification-not-removing-from-notification-center (#309)
* feature/CP-9383-ios-remove-notification-not-removing-from-notification-center

Added an optional argument for removing notification center from the notification center in the `removeNotification` function.

* feature/CP-9383-ios-remove-notification-not-removing-from-notification-center

Added an optional argument for removing notification center from the notification center in the `removeNotification` function.

* feature/CP-9383-ios-remove-notification-not-removing-from-notification-center

Added an optional argument for removing notification center from the notification center in the `removeNotification` function.

* Revert "feature/CP-9383-ios-remove-notification-not-removing-from-notification-center"

This reverts commit 96265879829a0c95ba58c4b13d7ad50c2b2d2598.

* Revert "feature/CP-9383-ios-remove-notification-not-removing-from-notification-center"

This reverts commit 6fd17bde1fdc9eef0452feefe825304e96eb8799.

* feature/CP-9383-ios-remove-notification-not-removing-from-notification-center

Added an optional argument for removing notification center from the notification center in the `removeNotification` function.

* feature/CP-9383-ios-remove-notification-not-removing-from-notification-center

Added an optional argument for removing notification center from the notification center in the `removeNotification` function.