feature/cp-9561-ios-mobile-sdk-call-the-api-created-to-communicate-and-track-v1 (#330)
* feature/cp-9561-ios-mobile-sdk-call-the-api-created-to-communicate-and-track-v1

Optimized API call to track when a user clicks a notification in the InboxView.

* feature/cp-9561-ios-mobile-sdk-call-the-api-created-to-communicate-and-track-v1

Optimized API call to track when a user clicks a notification in the InboxView.

* feature/cp-9561-ios-mobile-sdk-call-the-api-created-to-communicate-and-track-v1

Optimized API call to track when a user clicks a notification in the InboxView.

* feature/cp-9561-ios-mobile-sdk-call-the-api-created-to-communicate-and-track-v1

Optimized API call to track when a user clicks a notification in the InboxView.